+++ src/utils/RatingComparator.java
@@ -7,7 +7,7 @@ import backend.Place;
 public class RatingComparator<T extends Place> implements Comparator<T> {
 
 	public int compare(T o1, T o2) {
-		return ((new Integer(o1.getRating())).compareTo((new Integer(o2.getRating()))));
+		return ((new Double(o1.getRating())).compareTo((new Double(o2.getRating()))));
 	}
 
 }
